move oslib/directory.h function defs into a .cpp file
